fix(lint): 修复 ruff 0.16 规则漂移导致的 33 个报错，并显式固定 lint 规则集
根因：CI 的矩阵 job 按 `ruff>=0.5` 安装，实际装到 **0.16.7**，而本机是 0.14.4；
新版默认规则集更宽（新增 I / UP / DTZ / RUF / S / PLW / TRY 等），于是 6 个平台任务
全部卡在 Lint 步骤（33 errors；依赖锁定与 SARIF 冒烟两个 job 正常）。

固定工具链：
- pyproject：新增 `[tool.ruff.lint] select = ["E4","E7","E9","F","I","UP","DTZ"]`，
  只跑我们验证过的规则集，不再随 ruff 默认集合漂移；
  `ignore = ["UP042"]`（`str, Enum` → `StrEnum` 会改变 `str(Enum)` 的运行时返回值，单独评估）；
  dev 依赖把 ruff 钉到 `>=0.16,<0.17`；
- requirements-lock.txt：ruff 更新为已验证的 0.16.7。

修复内容（33 项）：
- DTZ：`core/audit_schedule.py` 两处 `datetime.now()` → `datetime.now(UTC)`；
  `tests/test_e2_e3.py` 的 cron 用例补 `tzinfo=UTC`；
- 自动修复 28 处：导入排序（I001）、`typing.Sequence/Iterable` → `collections.abc`（UP035）、
  `__all__` 排序（RUF022）、`datetime.timezone.utc` → `datetime.UTC`（UP017）；
- `int(round(...))` 冗余转换去掉（RUF046 ×4：cli / benchmark / stats）；
- 进度回调异常不再静默 `pass`，改为 `logger.debug(..., exc_info=True)`
  （`orchestrator/plan_execute.py`、`orchestrator/react_loop.py`）；
- `SandboxExecutor` 显式 `subprocess.run(..., check=False)`（PLW1510）；
- `webapp.py` 用 `TYPE_CHECKING` 导入 `AuditTargetSpec` 并去掉引号注解（UP037，保留懒加载意图）。

行为变更（次要）：规划器输出不是 JSON 对象时改抛 `TypeError`（替代 `ValueError`，ruff TRY004 建议）；
已确认无调用方依赖该异常类型，且便于区分"参数类型错误"与"解析失败"。

验证：ruff 0.16.7（CI 同款）与 0.14.4（本机）均 All checks passed；
pytest -q → 325 passed, 3 skipped。

+++ agent_shield/core/stats.py
@@ -88,7 +88,7 @@ def rate_estimate(
 
 def format_rate(estimate: RateEstimate, *, digits: int = 0) -> str:
     """人类可读：``100% (95% CI 44%–100%)``。"""
-    ci = int(round(estimate.confidence * 100))
+    ci = round(estimate.confidence * 100)
     return (
         f"{estimate.rate:.{digits}%} "
         f"({ci}% CI {estimate.ci_low:.{digits}%}–{estimate.ci_high:.{digits}%})"
